Changement de
lat=location.getLong("lat");
lng=location.getLong("lng");
				
				par
				
lat=location.getDouble("lat");
lng=location.getDouble("lng");

Permet d'avoir des résultats plus prècis ;)

+++ trunk/EyeWay/src/com/example/eyeway/fouilleDedonne/FouilleDonnee.java
@@ -142,8 +142,9 @@ public class FouilleDonnee {
 
 				//Position GPS
 				location=result.getJSONObject("geometry").getJSONObject("location");
-				lat=location.getLong("lat");
-				lng=location.getLong("lng");
+				
+				lat=location.getDouble("lat");
+				lng=location.getDouble("lng");
 
 				//L'icone
 				if(result.has("icon")){
